added a pop up upon delete account for confirmation

+++ client/src/components/UserDetails.js
@@ -6,11 +6,17 @@ function UserDetails() {
     const { currentUser, handleDeleteAccount } = useContext(UserContext)
     const [isHidden, setIsHidden] = useState(true)
 
+    function confirmDelete(){
+        if (window.confirm("Are you sure you want to delete your account? This action cannot be undone.")) {
+            handleDeleteAccount()
+        }
+    }
+
     return (
         <div>
             <button onClick={() => setIsHidden(!isHidden)}>Edit My Info</button>
             <button onClick={() => {
-                handleDeleteAccount()
+                confirmDelete()
                 setIsHidden(true)}}>Delete My Account</button>
             {isHidden ? null : <EditProfileForm/>}
         </div>
